fazendo o relacionamento entre tabelas

+++ src/postagem/services/postagem.service.ts
@@ -1,53 +1,86 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
-import { InjectRepository } from '@nestjs/typeorm';
-import { ILike, Repository } from 'typeorm';
+import { DeleteResult, ILike, Repository } from 'typeorm';
 import { Postagem } from '../entities/postagem.entity';
-import { DeleteResult } from 'typeorm/browser';
+import { InjectRepository } from '@nestjs/typeorm';
+import { TemaService } from '../../tema/services/tema.service';
 
 @Injectable()
 export class PostagemService {
   constructor(
     @InjectRepository(Postagem)
     private postagemRepository: Repository<Postagem>,
+    private temaService: TemaService,
   ) {}
+
   async findAll(): Promise<Postagem[]> {
-    return this.postagemRepository.find();
+    return await this.postagemRepository.find({
+      relations: {
+        tema: true,
+      },
+    });
   }
-  async findByID(id: number): Promise<Postagem> {
+
+  async findById(id: number): Promise<Postagem> {
     const postagem = await this.postagemRepository.findOne({
       where: { id },
+      relations: {
+        tema: true,
+      },
     });
+
     if (!postagem) {
-      throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
+      throw new HttpException('Postagem não encontrada', HttpStatus.NOT_FOUND);
     }
+
     return postagem;
   }
-  async findByTitle(title: string): Promise<Postagem[]> {
+
+  async findByTitulo(titulo: string): Promise<Postagem[]> {
     return await this.postagemRepository.find({
       where: {
-        titulo: ILike(`%${title}%`),
+        titulo: ILike(`%${titulo}%`),
+      },
+      relations: {
+        tema: true,
       },
     });
   }
 
   async create(postagem: Postagem): Promise<Postagem> {
-    return this.postagemRepository.save(postagem);
+    if (postagem.tema) {
+      const tema = await this.temaService.findById(postagem.tema.id);
+
+      if (!tema) {
+        throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+      }
+    }
+
+    return await this.postagemRepository.save(postagem);
   }
 
   async update(postagem: Postagem): Promise<Postagem> {
-    const postAtualizada = postagem.id;
+    const buscaPostagem = await this.findById(postagem.id);
 
-    if (!postAtualizada) {
-      throw new HttpException(
-        'Postagem não encontrada',
-        HttpStatus.BAD_REQUEST,
-      );
+    if (!buscaPostagem || !postagem.id) {
+      throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
     }
-    return this.postagemRepository.save(postagem);
+
+    if (postagem.tema) {
+      const tema = await this.temaService.findById(postagem.tema.id);
+
+      if (!tema) {
+        throw new HttpException('Tema não encontrado!', HttpStatus.NOT_FOUND);
+      }
+    }
+
+    return await this.postagemRepository.save(postagem);
   }
 
   async delete(id: number): Promise<DeleteResult> {
-    await this.findByID(id);
+    const buscaPostagem = await this.findById(id);
+
+    if (!buscaPostagem)
+      throw new HttpException('Postagem não encontrada!', HttpStatus.NOT_FOUND);
 
     return await this.postagemRepository.delete(id);
   }
